+++ cmd/hoarctl/store.go
@@ -6,8 +6,8 @@ import (
 	"os"
 
 	cli "github.com/jawher/mow.cli"
-	"github.com/monax/hoard"
-	"github.com/monax/hoard/reference"
+	"github.com/monax/hoard/v3"
+	"github.com/monax/hoard/v3/reference"
 )
 
 // Cat retrieves encrypted data from store