+++ config/storage/filesystem.go
@@ -5,7 +5,7 @@ import (
 	"path"
 
 	"github.com/cep21/xdgbasedir"
-	"github.com/monax/hoard/storage"
+	"github.com/monax/hoard/v3/storage"
 )
 
 type FileSystemConfig struct {