+++ logging/convention.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"github.com/go-kit/kit/log"
-	"github.com/monax/hoard/logging/structure"
+	"github.com/monax/hoard/v3/logging/structure"
 )
 
 func InfoLogger(logger log.Logger) log.Logger {