+++ project/history.go
@@ -31,6 +31,13 @@ func FullVersion() string {
 var History relic.ImmutableHistory = relic.NewHistory("Monax Hoard", "https://github.com/monax/hoard").
 	MustDeclareReleases("",
 		``,
+		"3.2.0 - 2019-04-23",
+		`### Fixed
+- [MODULES] Add v3 to module declaration and update imports
+
+### Removed
+- [BUILD] Remove vendor and related scripting
+`,
 		"3.1.0 - 2019-04-23",
 		`### Added
 - [SERVER] Added Wait() function to wait until server is ready and ListenAddress for getting bound listen address (useful when using localhost:0 for a OS selected free port)