+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 ## [Unreleased]
 
 
+## [3.2.0] - 2019-04-23
+### Fixed
+- [MODULES] Add v3 to module declaration and update imports
+
+### Removed
+- [BUILD] Remove vendor and related scripting
+
+
 ## [3.1.0] - 2019-04-23
 ### Added
 - [SERVER] Added Wait() function to wait until server is ready and ListenAddress for getting bound listen address (useful when using localhost:0 for a OS selected free port)
@@ -92,7 +100,8 @@ This is the first Hoard open source release and includes:
 	- Hoar-Daemon hoard
 	- Hoar-Control hoarctl CLI
 
-[Unreleased]: https://github.com/monax/hoard/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/monax/hoard/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/monax/hoard/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/monax/hoard/compare/v3.0.1...v3.1.0
 [3.0.1]: https://github.com/monax/hoard/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/monax/hoard/compare/v2.0.0...v3.0.0